Converted client login start packet
