+++ src/main/kotlin/packets/PacketHandler.kt
@@ -100,9 +100,6 @@ import kotlin.time.Duration.Companion.seconds
 class PacketHandler(
     private val client: ClientSession
 ) {
-    val world: World
-        get() = client.player.world!!
-
     @PacketReceiver
     fun onEncryptionResponse(packet: ClientEncryptionResponsePacket) {
         val rsa = Cipher.getInstance("RSA/ECB/PKCS1Padding")
@@ -175,48 +172,6 @@ class PacketHandler(
     }
 
     /**
-     * Handles when the client tells the server it's ready to log in
-     *
-     * The server first checks for a valid version and uuid, then sends a login success packet
-     * It'll then transition the game state into play mode
-     * and send a join game and player position/look packet to get past all loading screens
-     */
-    @PacketReceiver
-    fun onLoginStart(packet: ClientLoginStartPacket) {
-        val preJoinEvent = PlayerPreJoinEvent()
-        EventManager.fire(preJoinEvent)
-        if (preJoinEvent.isCancelled) return
-
-        val username = packet.username
-        val uuid = UUID.nameUUIDFromBytes(("OfflinePlayer:$username").toByteArray())
-
-        checkLoginValidity(username)
-
-        val player = initializePlayer(username, uuid)
-        if (!Bullet.onlineMode) {
-            val dupes = players.filter {
-                it.username == username || it.uuid == uuid
-            }
-
-            players.removeAll(dupes)
-            dupes.forEach { old ->
-                old.sendPacket(
-                    ServerLoginDisconnectPacket(
-                        Component.text("You are already logged in from another location", NamedTextColor.RED)
-                    )
-                )
-
-                old.clientSession.close()
-            }
-
-            players.add(player)
-            loginPlayer()
-        }
-
-        handleOnlineModeJoin(packet)
-    }
-
-    /**
      * Handles a ping packet by sending a pong response and closing the connection
      */
     @PacketReceiver
@@ -272,268 +227,6 @@ class PacketHandler(
         }
     }
 
-    private fun initializePlayer(username: String, uuid: UUID): Player {
-        val player = Player(client)
-        player.username = username
-        player.uuid = uuid
-
-        player.location = LocationType.Location(8.5, 2.0, 8.5)
-        player.onGround = false
-
-        if (player.gameMode != GameMode.SURVIVAL || player.gameMode != GameMode.ADVENTURE) {
-            player.canFly = true
-        }
-
-        client.player = player
-        return player
-    }
-
-    private fun sendSpawnPlayerPackets(player: Player) {
-        for (otherPlayer in Bullet.players) {
-            if (otherPlayer != player) {
-                otherPlayer.clientSession.sendPacket(
-                    ServerSpawnPlayerPacket(
-                        player.entityID,
-                        player.uuid,
-                        player.location
-                    )
-                )
-            }
-        }
-
-        for (existingPlayer in Bullet.players) {
-            if (existingPlayer != player) {
-                client.sendPacket(
-                    ServerSpawnPlayerPacket(
-                        existingPlayer.entityID,
-                        existingPlayer.uuid,
-                        existingPlayer.location
-                    )
-                )
-            }
-        }
-    }
-
-    private fun checkLoginValidity(username: String): Boolean {
-        if (client.protocol > Bullet.PROTOCOL) {
-            client.sendPacket(
-                ServerLoginDisconnectPacket(
-                    Component.text()
-                        .append(Component.text("Your client is outdated, please downgrade to minecraft version"))
-                        .append(Component.text(" " + Bullet.VERSION).color(NamedTextColor.GOLD))
-                        .build()
-                )
-            )
-
-            client.close()
-            return false
-        } else if (client.protocol < Bullet.PROTOCOL) {
-            client.sendPacket(
-                ServerLoginDisconnectPacket(
-                    Component.text()
-                        .append(Component.text("Your client is outdated, please upgrade to minecraft version"))
-                        .append(Component.text(" " + Bullet.VERSION).color(NamedTextColor.GOLD))
-                        .build()
-                )
-            )
-
-            return false
-        }
-
-        if (!username.matches(Regex("^[a-zA-Z0-9]{3,16}$"))) {
-            client.sendPacket(ServerLoginDisconnectPacket(Component.text("Invalid username")))
-            return false
-        }
-
-        return true
-    }
-
-    private fun scheduleTimers() {
-        client.sendPlayerSpawnPacket()
-        client.scheduleKeepAlive()
-        client.scheduleHalfSecondUpdate()
-
-        if (Bullet.shouldPersist) client.scheduleSaving()
-    }
-
-    private fun readPlayerPersistentData() {
-        val player = client.player
-        val data = Bullet.storage.storage.readPlayerData(player.uuid) ?: return
-        player.status.health = data.health
-        player.status.foodLevel = data.foodLevel
-        player.status.saturation = data.saturation
-        player.status.exhaustion = data.exhaustionLevel
-        player.location = data.location
-        player.permissionLevel = data.permissionLevel
-        player.totalXP = data.totalXP
-
-        player.setGameMode(GameMode.entries.find { it.id == data.gameMode } ?: GameMode.SURVIVAL)
-
-        val savedStacks: Map<Int, ItemStack> =
-            data.inventory.associate { it.first to it.second }
-        player.inventory.clear()
-
-        val totalSlots = 45
-        val slotDataList = (0 until totalSlots).map { idx ->
-            val stack = savedStacks[idx]
-            player.inventory.set(idx, stack)
-            stack?.toSlotData() ?: Slot.SlotData(false)
-        }
-
-        player.sendPacket(ServerWindowItemsPacket(0, slotDataList))
-        sendHeldItemUpdate()
-        calculateXPLevels(player.totalXP)
-
-        player.sendPacket(
-            ServerUpdateHealthPacket(
-                player.status.health.toFloat(),
-                player.status.foodLevel,
-                player.status.saturation
-            )
-        )
-
-        player.sendPacket(ServerPlayerPositionAndLookPacket(player.location))
-
-    }
-
-    private fun sendBlockChanges() {
-        if (world.modifiedBlocks.isEmpty()) return
-        for ((pos, meta) in world.modifiedBlocks) {
-            client.player.sendPacket(
-                ServerBlockChangePacket(
-                    pos, meta.stateID
-                )
-            )
-
-            if (meta.textLines != null) {
-                val nbt = CompoundTag()
-                nbt.putString("id", "minecraft:sign")
-                nbt.putInt("x", pos.x.toInt())
-                nbt.putInt("y", pos.y.toInt())
-                nbt.putInt("z", pos.z.toInt())
-
-                meta.textLines.forEachIndexed { idx, line ->
-                    nbt.putString("Text${idx + 1}", "{\"text\":\"$line\"}")
-                }
-
-                client.player.sendPacket(
-                    ServerBlockEntityDataPacket(
-                        pos, 9, nbt
-                    )
-                )
-            }
-        }
-    }
-
-    private fun checkForBan(): Boolean {
-        // Get ban or return true i
-        val ban = Bullet.storage.getPlayerBan(client.player.uuid) ?: return false
-
-        val durationMillis = ban.duration.inWholeMilliseconds
-        val banEnd = ban.currentTime + durationMillis
-        val now = System.currentTimeMillis()
-
-        if (durationMillis > 0 && now >= banEnd) {
-            Bullet.storage.unbanPlayer(ban.uuid)
-            return false
-        }
-
-        val expirationText = if (durationMillis <= 0) {
-            "permanently"
-        } else {
-            val expirationTime = Instant.ofEpochMilli(banEnd)
-                .atZone(ZoneId.systemDefault())
-
-            val dayOfMonth = expirationTime.dayOfMonth
-            val daySuffix = DurationFormat.getDaySuffix(dayOfMonth)
-
-            val formattedDate = expirationTime.format(
-                DateTimeFormatter.ofPattern("MMMM d'$daySuffix' yyyy 'at' H:mm")
-            )
-
-            "Expires $formattedDate"
-        }
-
-        client.disconnect(
-            Component.text()
-                .append(Component.text("You have been banned!\n", NamedTextColor.RED))
-                .append(Component.text(expirationText, NamedTextColor.RED))
-                .append(Component.text("\n\n", NamedTextColor.RED))
-                .append(Component.text("Reason: ", NamedTextColor.RED))
-                .append(Component.text(ban.reason, NamedTextColor.GRAY))
-                .build()
-        )
-
-        return true
-    }
-
-    private fun sendEntities() {
-        for (entity in world.entities) {
-            client.sendPacket(
-                ServerSpawnEntityPacket(
-                    entity.first.entityID,
-                    entity.second.uuid,
-                    entity.second.entityType,
-                    entity.second.location,
-                    entity.second.velocityX,
-                    entity.second.velocityY,
-                    entity.second.velocityZ,
-                )
-            )
-        }
-
-        for (livingEntity in world.livingEntities) {
-            client.sendPacket(
-                ServerSpawnLivingEntityPacket(
-                    livingEntity.first.entityID,
-                    livingEntity.second.uuid,
-                    livingEntity.second.entityType,
-                    livingEntity.second.location,
-                    livingEntity.second.headPitch,
-                    livingEntity.second.velocityX,
-                    livingEntity.second.velocityY,
-                    livingEntity.second.velocityZ,
-                )
-            )
-        }
-    }
-
-    fun handleOnlineModeJoin(packet: ClientLoginStartPacket) {
-        if (Bullet.onlineMode) {
-            val verifyToken = ByteArray(4).apply {
-                SecureRandom().nextBytes(this)
-            }
-
-            client.verifyToken = verifyToken
-            client.player.sendPacket(
-                ServerEncryptionRequestPacket(
-                    "",
-                    Bullet.publicKey,
-                    verifyToken
-                )
-            )
-        }
-    }
-
-    private fun sendJoinGamePacket() {
-        val player = client.player
-        client.sendPacket(
-            ServerJoinGamePacket(
-                player.entityID,
-                false,
-                player.gameMode,
-                "minecraft:overworld",
-                Bullet.dimensionCodec!!,
-                Bullet.max_players,
-                32,
-                reducedDebugInfo = false,
-                enableRespawnScreen = true,
-                isDebug = false,
-                isFlat = true
-            )
-        )
-    }
-
     private fun verifyPlayerToken(verifyToken: ByteArray) {
         if (!client.verifyToken.contentEquals(verifyToken)) {
             client.sendPacket(
@@ -546,43 +239,5 @@ class PacketHandler(
             client.close()
             return
         }
-    }
-
-    private fun loginPlayer() {
-        val player = client.player
-        client.sendPacket(ServerLoginSuccessPacket(player.uuid, player.username))
-        client.state = GameState.PLAY
-
-        if(checkForBan()) return
-
-        sendJoinGamePacket()
-        client.sendPacket(ServerPlayerPositionAndLookPacket(player.location))
-
-        players.add(player)
-        readPlayerPersistentData()
-        scheduleTimers()
-
-        client.sendPacket(ServerChunkPacket(0, 0))
-        sendSpawnPlayerPackets(player)
-
-        client.sendPacket(ServerUpdateViewPositionPacket(player.chunkX, player.chunkZ))
-        client.updatePlayerChunks(player.chunkX, player.chunkZ)
-
-        sendBlockChanges()
-        sendEntities()
-
-        Bullet.storage.storage.writePlayerData(player)
-
-        val joinEvent = PlayerJoinEvent(client.player)
-        EventManager.fire(joinEvent)
-        if (joinEvent.isCancelled) return
-
-        val world = player.world!!
-        player.setTimeOfDay(world.timeOfDay)
-        if (world.weather == 1) player.sendPacket(ServerChangeGameStatePacket(2, 0f))
-        else player.sendPacket(ServerChangeGameStatePacket(1, 0f))
-
-        val (nodes, rootIndex) = buildCommandGraphFromDispatcher(CommandManager.dispatcher)
-        client.sendPacket(ServerDeclareCommandsPacket(nodes, rootIndex))
     }
 }
